Add `Debug` bounds to iterator and factory; expand macro docs.
This is a significant choice. I believe it is better to require them
than not to, because this allows the derived iterator and factory to
implement `Debug` to actually print details, whereas if we did not,
we would have to either wait for RFC 2056 `trivial_bounds` or require
the user of the macro to explicitly specify “also generate Debug”.

These bounds do not affect what types can implement `Exhaust`, other
than if `type Factory = Self;` is used and `Self` does not implement
`Debug`, a wrapper may be needed.

The derive macro now implements `Debug` for the factory.