Change macro-generated `::core` paths to a helper module.
This way, the macro depends on exactly one crate being as expected,
and if we ever get $crate for proc macros, or add a "set your own path"
configuration attribute to the macro, we can plug it right in.

It also gives a better place to put `ExhaustWithFactoryEqSelf` and any
future such items.