Add special case when deriving for 1 field.
In this case, we don’t need peekable iterators, and can just delegate,
saving the memory for the peeked value.

Note that, unlike the internal `delegate_factory_and_iter!()`, this does
not delegate the whole iterator type, but still always wraps it. Doing
that will require an explicit configuration by the user.