Remove superfluous parentheses from tuple impls.
This has no significant effect, but it's 20 fewer tokens for rustc to
process, so why not. Previously, the macro would generate code like

    let factory = ((crate::mh_::clone(factory_0), factory_1));

and with this change it is

    let factory = (crate::mh_::clone(factory_0), factory_1);

without the parentheses that would be a tuple-struct for the public
factory type if there was a public factory type.

This was discovered by a spurious warning from Rust 1.92 beta
<https://github.com/rust-lang/rust-clippy/issues/16023>, so this will
also prevent that warning from showing up in CI until it is fixed by
Rust beta backports.